Edited value types from BiometricFacial and BiometricFingerPrint to BigIntegers.
Added an util function to check whether BigIntegers are similar or not (< 1% of increase between them).

Added an alternative method into BiometricVerification in order to use it if it's not null and biometricverification fails to get verify biometric data or is unable to get dni from passport.

Edited all the tests to satisfy the new atributes from the edited value-classes.

Added an BigInteger commented in Main that not satisfies the < 1% of increase and fails the biometric method in order to see if alternative is executed correctly.

This should be the last commit of the project ;)